Move LifecycleScopes, Lifecycle boundary exceptions to the new artifact
