+++ sample/src/main/java/com/uber/autodispose/recipes/AutoDisposeActivity.java
@@ -19,7 +19,7 @@ package com.uber.autodispose.recipes;
 import android.app.Activity;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
-import com.uber.autodispose.LifecycleEndedException;
+import com.uber.autodispose.lifecycle.LifecycleEndedException;
 import com.uber.autodispose.lifecycle.LifecycleScopeProvider;
 import io.reactivex.Observable;
 import io.reactivex.functions.Function;
